add many features: save file, normal map and fix many bugs

+++ Opengl/Objects/Item.cpp
@@ -1,13 +1,13 @@
 #include "Item.h"
 
-Item::Item(long id, PointXYZ position)
+Item::Item(long long id, PointXYZ position)
 	: DynamicObject(id, position)
 {
 	this->name = "Item";
 	this->interactable = true;
 }
 
-Item::Item(long id)
+Item::Item(long long id)
 	: Item(id, PointXYZ(0,0,0))
 {
 }
